Refactored config validation out of configure_cse to config_validator
