+++ container_service_extension/cse.py
@@ -13,8 +13,8 @@ from pyVmomi import vim
 import requests
 from vcd_cli.utils import stdout
 
+from container_service_extension.config_validator import get_validated_config
 from container_service_extension.configure_cse import check_cse_installation
-from container_service_extension.configure_cse import get_validated_config
 from container_service_extension.configure_cse import install_cse
 from container_service_extension.exceptions import AmqpConnectionError
 from container_service_extension.sample_generator import generate_sample_config
